refactor(journal): repository + service classes; consolidate lib/journals.ts
Applies the same architectural pattern established for templates: a
JournalRepository class for file-level CRUD over the user's journal,
and a JournalService class on top for business logic (validation,
fingerprint guards, mutex, cache invalidation). The standalone
lib/journals.ts, lib/journal/write.ts, and lib/journal/backfill.ts
modules are absorbed into the class structure.

Data layer:
- lib/journal/layout.ts — DEFAULT_MAIN, PRICE_DB_NAME, VALID_EXTS,
  getJournalDir, getJournalCacheTag. Pure helpers; reads DATA_DIR
  from process.env lazily so the module stays free of env validation.
- lib/journal/repository.ts — JournalRepository class with find,
  list, getLayout, ensureLayout, setMainFile, emptyJournalDir,
  readFile, writeFileAtomic, appendFile. Constructor takes a
  DbInstance for testability.
- lib/journal/service.ts — JournalService class with addTransaction,
  editTransaction, deleteTransaction (mutex + fingerprint), backfillUids,
  replaceFromSingleFile, replaceFromZip, plus passthrough list/find
  helpers. Constructor takes a JournalRepository.
- lib/journal/index.ts — production singletons (journalRepository,
  journalService) bound to the default db, plus type re-exports.

Callers updated:
- features/transactions/actions/{create,update,delete}Transaction.ts
  → use journalService methods.
- features/transactions/Transactions.tsx → journalService.listTransactions.
- features/transactions/EditTransaction.tsx → journalService.findTransaction.
- app/api/upload/route.ts → journalService.replaceFromSingleFile /
  replaceFromZip.
- utils/runLedger.ts → journalRepository.ensureLayout +
  getJournalCacheTag from layout.

Tests:
- lib/journal/service.test.ts consolidates the old write.test.ts,
  backfill.test.ts, and journals.test.ts.
- lib/journal/integration.test.ts switches to class instances.
- lib/templates/integration.test.ts uses JournalService for the
  "use a template" flow.

Removed: lib/journals.ts, lib/journal/write.ts, lib/journal/backfill.ts,
lib/journals.test.ts, lib/journal/write.test.ts,
lib/journal/backfill.test.ts.

84 tests pass. Type-check + lint clean.
