+++ features/transactions/actions/createTransaction.ts
@@ -2,8 +2,7 @@
 
 import type { TransactionActionState } from './types';
 import { requireUser } from '@/lib/auth/require-user';
-import { addTransaction, getJournalCacheTag } from '@/lib/journals';
-import { revalidatePath, updateTag } from 'next/cache';
+import { journalService } from '@/lib/journal';
 
 export async function createTransactionAction(
   _prev: TransactionActionState | null,
@@ -22,7 +21,7 @@ export async function createTransactionAction(
     return { ok: false, formError: 'Transaction payload is not valid JSON' };
   }
 
-  const result = await addTransaction(user.id, parsed);
+  const result = await journalService.addTransaction(user.id, parsed);
   if (!result.ok) {
     return {
       ok: false,
@@ -30,8 +29,5 @@ export async function createTransactionAction(
       formError: result.formError,
     };
   }
-
-  updateTag(getJournalCacheTag(user.id));
-  revalidatePath('/', 'layout');
   return { ok: true };
 }
